fix: rename labels in Add New User form and update property names in user list
